// frontend/src/app/api/model/createReleaseStageInput.ts
/**
 * Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateReleaseStageInput { 
    stageId: number;
    stageName?: string | null;
    workingDays?: number;
}